requirements.txt
loguru
hydra-core
pyyaml
boto3
requests